feat(auth): add i18n keys for sign-in redesign (ARC-690)
Split apps/web/src/shared/i18n/messages/auth.ts into per-locale files
under messages/auth/ to stay under the 500-line file limit, and add
form, brand, pwa, and provider keys for the redesigned auth page.

(ARC-690)
