+++ apps/web/src/shared/i18n/messages/auth/index.ts
@@ -1,0 +1,19 @@
+import type { DeepPartial } from '../../base-types';
+import { en } from './en';
+import { es } from './es';
+import { fr } from './fr';
+import { ru } from './ru';
+import { by } from './by';
+
+export { en, es, fr, ru, by };
+
+export const authMessages = {
+  en,
+  es,
+  fr,
+  ru,
+  by,
+} as const;
+
+/** Derived type with Partial wrapper for backward compatibility */
+export type AuthMessages = DeepPartial<typeof en>;
